# server/api/openapi.gen.yaml
...
components:
    parameters:
        DocListCursor:
            in: query
            name: cursor
            required: false
            schema:
                type: string
        DocListDateFrom:
            in: query
            name: dateFrom
            required: false
            schema:
                format: date
                type: string
        DocListDateTo:
            in: query
            name: dateTo
            required: false
            schema:
                format: date
                type: string
        DocListLimit:
            in: query
            name: limit
            required: false
            schema:
                default: 50
                type: integer
        DocListQ:
            in: query
            name: q
            required: false
            schema:
                type: string
        DocListStatus:
            in: query
            name: status
            required: false
            schema:
                type: string
        DocListWarehouseId:
            in: query
            name: warehouseId
            required: false
            schema:
                format: uuid
                type: string
        DocumentId:
            in: path
            name: id
            required: true
            schema:
                format: uuid
                type: string
...
    responses:
        Conflict:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Resource already exists
        Forbidden:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Authenticated but not permitted
        NotFound:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Resource not found
...
        Unauthorized:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Not authenticated
        ValidationError:
            content:
                application/json:
                    schema:
                        $ref: '#/components/schemas/Error'
            description: Request failed validation
    schemas:
...
        CancelInput:
            properties:
                reason:
                    minLength: 1
                    type: string
            required:
                - reason
            type: object
...
        DecimalString:
            description: A decimal quantity or money value, string-encoded to avoid float loss
            example: "12.500"
            type: string
...
        DocumentActor:
            properties:
                displayName:
                    type: string
                id:
                    format: uuid
                    type: string
            required:
                - id
                - displayName
            type: object
        DocumentStatus:
            description: draft is editable; posted is immutable; reversed was cancelled by a reversal; cancelled is terminal for PO/SO (ADR-0018)
            enum:
                - draft
                - posted
                - reversed
                - cancelled
            type: string
        Error:
            properties:
                code:
                    type: string
                fields:
                    additionalProperties:
                        type: string
                    description: Per-field validation messages, when applicable
                    type: object
                message:
                    type: string
            required:
                - code
                - message
            type: object
        FulfillmentState:
            description: Server-computed per PO/SO line and document (ADR-0016); null until SN-0002 implements the rollup
            enum:
                - open
                - partial
                - closed
            type: string
...
        PurchaseOrder:
            properties:
                cancelReason:
                    type: string
                cancelledAt:
                    format: date-time
                    type: string
                createdAt:
                    format: date-time
                    type: string
                createdBy:
                    $ref: '#/components/schemas/DocumentActor'
                docDate:
                    format: date
                    type: string
                docNumber:
                    nullable: true
                    type: string
                fulfillmentState:
                    $ref: '#/components/schemas/FulfillmentState'
                id:
                    format: uuid
                    type: string
                lines:
                    items:
                        $ref: '#/components/schemas/PurchaseOrderLine'
                    type: array
                notes:
                    type: string
                postedAt:
                    format: date-time
                    type: string
                postedBy:
                    $ref: '#/components/schemas/DocumentActor'
                reversedById:
                    format: uuid
                    nullable: true
                    type: string
                reversesId:
                    format: uuid
                    nullable: true
                    type: string
                status:
                    $ref: '#/components/schemas/DocumentStatus'
                supplierId:
                    format: uuid
                    type: string
                warehouseId:
                    format: uuid
                    type: string
            required:
                - id
                - status
                - supplierId
                - warehouseId
                - docDate
                - notes
                - lines
                - createdAt
                - createdBy
            type: object
        PurchaseOrderInput:
            properties:
                docDate:
                    format: date
                    type: string
                lines:
                    items:
                        $ref: '#/components/schemas/PurchaseOrderLineInput'
                    minItems: 1
                    type: array
                notes:
                    type: string
                supplierId:
                    format: uuid
                    type: string
                warehouseId:
                    format: uuid
                    type: string
            required:
                - supplierId
                - warehouseId
                - docDate
                - lines
            type: object
        PurchaseOrderLine:
            properties:
                fulfillmentState:
                    $ref: '#/components/schemas/FulfillmentState'
                id:
                    format: uuid
                    type: string
                lineNo:
                    type: integer
                productId:
                    format: uuid
                    type: string
                qty:
                    $ref: '#/components/schemas/DecimalString'
                receivedQty:
                    $ref: '#/components/schemas/DecimalString'
                unitCost:
                    $ref: '#/components/schemas/DecimalString'
                uom:
                    type: string
            required:
                - id
                - lineNo
                - productId
                - uom
                - qty
                - unitCost
            type: object
        PurchaseOrderLineInput:
            properties:
                productId:
                    format: uuid
                    type: string
                qty:
                    $ref: '#/components/schemas/DecimalString'
                unitCost:
                    $ref: '#/components/schemas/DecimalString'
                uom:
                    minLength: 1
                    type: string
            required:
                - productId
                - uom
                - qty
            type: object
        PurchaseOrderList:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/PurchaseOrder'
                    type: array
                nextCursor:
                    nullable: true
                    type: string
            required:
                - items
                - nextCursor
            type: object
...
    securitySchemes:
        sessionCookie:
            in: cookie
            name: sinta_session
            type: apiKey
info:
    description: |
        Multi-tenant inventory management for Indonesian SME distributors. Contract-first: this file is the single source of truth for both the Go server (oapi-codegen) and the TypeScript client (openapi-typescript). Conventions: camelCase fields, kebab-case plural paths, /v1 prefix (docs/reference/CONVENTIONS.md).
    title: Sinta API
    version: 0.1.0
openapi: 3.0.3
paths:
...
    /purchase-orders:
        get:
            operationId: listPurchaseOrders
            parameters:
                - $ref: '#/components/parameters/DocListStatus'
                - $ref: '#/components/parameters/DocListWarehouseId'
                - $ref: '#/components/parameters/DocListDateFrom'
                - $ref: '#/components/parameters/DocListDateTo'
                - $ref: '#/components/parameters/DocListQ'
                - $ref: '#/components/parameters/DocListCursor'
                - $ref: '#/components/parameters/DocListLimit'
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PurchaseOrderList'
                    description: Purchase orders
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
            summary: List purchase orders
            tags:
                - documents
        post:
            operationId: createPurchaseOrder
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PurchaseOrderInput'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PurchaseOrder'
                    description: Draft created
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "422":
                    $ref: '#/components/responses/ValidationError'
            summary: Create a draft purchase order
            tags:
                - documents
    /purchase-orders/{id}:
        delete:
            operationId: deletePurchaseOrder
            responses:
                "204":
                    description: Deleted
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
                "409":
                    $ref: '#/components/responses/Conflict'
            summary: Delete a draft purchase order (rejected once posted)
            tags:
                - documents
        get:
            operationId: getPurchaseOrder
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PurchaseOrder'
                    description: Purchase order
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
            summary: Get a purchase order
            tags:
                - documents
        parameters:
            - $ref: '#/components/parameters/DocumentId'
        put:
            operationId: updatePurchaseOrder
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/PurchaseOrderInput'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PurchaseOrder'
                    description: Draft updated
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
                "409":
                    $ref: '#/components/responses/Conflict'
                "422":
                    $ref: '#/components/responses/ValidationError'
            summary: Replace a draft purchase order (rejected once posted)
            tags:
                - documents
    /purchase-orders/{id}/cancel:
        parameters:
            - $ref: '#/components/parameters/DocumentId'
        post:
            operationId: cancelPurchaseOrder
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CancelInput'
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PurchaseOrder'
                    description: Cancelled
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
                "409":
                    $ref: '#/components/responses/Conflict'
            summary: Cancel a draft purchase order (terminal status, ADR-0018)
            tags:
                - documents
    /purchase-orders/{id}/post:
        parameters:
            - $ref: '#/components/parameters/DocumentId'
        post:
            operationId: postPurchaseOrder
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PurchaseOrder'
                    description: Posted
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
                "409":
                    $ref: '#/components/responses/Conflict'
            summary: Post a draft purchase order (assigns the gapless number)
            tags:
                - documents
    /purchase-orders/{id}/reverse:
        parameters:
            - $ref: '#/components/parameters/DocumentId'
        post:
            operationId: reversePurchaseOrder
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PurchaseOrder'
                    description: Reversal document posted
                "401":
                    $ref: '#/components/responses/Unauthorized'
                "403":
                    $ref: '#/components/responses/Forbidden'
                "404":
                    $ref: '#/components/responses/NotFound'
                "409":
                    $ref: '#/components/responses/Conflict'
            summary: Reverse a posted purchase order (posts a cancelling document)
            tags:
                - documents
...
security:
    - sessionCookie: []
servers:
    - url: /v1